Solutions of part 2 (Course Information) 2.5
